map unicode bindings to vk-s; add key names
Some applications reject SendUnicode input, so if the keyboard layout
(for now, assume the default is the only) has the Unicode character
being mapped, bind it as a vk-to-vk remap.

kl: add pressing ctrl/alt/shift for while sending the binding
os: add determining needed ctrl/alt/shift state from a Unicode char
    (Ctrl is needed because AltGr is LeftCtrl+RightAlt)

Add "kn.cpp" list of key names.
For it, also add "scancodes.h" defining "standard" keycodes (they are
said to be device-specific, but this subset seems to be standard; it is
probably multimedia, etc. keys that differ in practice).

This makes use of Windows-specific functions, and only builds under Mingw.