add vk-from-unicode caching
This way, determining vk and ctrl/alt/shift state corresponding to the
remapping in Windows layout only needs to be done on the first switching
to the layout it is defined for.

Name it "compiling"; "KLC" corresponds to "compiled KLY", e.g.
"compiled Keyboard Layout", e.g. "compiled set of remappings".