add modifier state tracking
This way, code that temprorary brings ctrl/alt/shift into a state
required by a binding is aware of their current state.
Before, if a modifier was already in effect, the bringing would still
send a press, and afterwards a release, cancelling the modifier.

Add "km.cpp" that implements modifier state tracker.
Call it from kl.cpp on press/release.
Adjust modifier actions in ka.cpp to use the tracker.