add per-window-title sc->sc binds
This feature is intended for games with limited or lacking keyboard
bindings customization.

ka:
  Add a dedicated toggle action.
  This is intended for entering in-game text (e.g. in chat)

eh:
  Call kr.cpp on window switch to match the title and apply the binds
  then.

kl:
  Re-add sc->sc sending, fixing it (a keybd_event flag was missing).
  The sc->sc to sc->vk translation for non-per-window-title binds is left
  as is for now, though.
  Add modifying directly the active KLY, for kr.cpp to use.

To avoid comparing the focused window title with every title defined for
remapping, require the window size to first match one of user-supplied
resolutions.  This way, non-"game" windows, where heavy typing happens,
would not see a delay in typing just after the window switch if the
nuber of window titles for which per-title remaps are defined is large.

Add a delay to let a minimized window animate its restore before
examining the dimensions.

parse:
  Add "title" directive for defining window titles.
  Add "res" directive for defining allowed resolutions.

main:
  Add example per-window-title bindings.

kr:
  Implements window title matching, keeping the per-window-title
  remaps, and modifying the active KLY accordingly.