add reading of for-when-ctrl/alt/win layout
The purpose of this feature, once implemented, is to make ctrl/alt/win
key combinations, such as Ctrl-Z, consistent across all layouts.

Explanation:
Keyboard shortcuts are usually defined in terms of VKs.
System layouts put VKs to correspond to Qwerty.
Remapping a key means moving a VK to another physical position.
A Latin remapping (qwaf) results in VKs being laid out one way,
and a non-Latin remapping (qwaf ru) results in another.
Using two sets of shortcuts is uncomfortable.
Therefore, make one layout be "always active when a modifier is in effect".

Add the vks_lang directive to set the "while modifiers are in effect" layout.
Implement it by adding a "vk lookup table for while modifiers are in effect",
the KL_mods_vks, which is not yet actually utilized by kl.cpp.
Default to LANG_NEUTRAL, which should select the first system layout.

main:
  Add setting the LANG_ENGLISH as the while-modifiers-are-in-effect-lang.

Also add an explicit binding of Space key to its character to fix it not
being present on level 5.