fix occasional cancellation of level5
When a modifier is down, a special when-modifier-is-in-effect VK-to-send
table is used.  This is so that Ctrl+Letter combinations do not differ
across layouts.

However, when the following happens:

  All keys are up
  Modifier down
  Level5 down
  Modifier up
  A letter down fast enough to be before another Level5 down

Then the Level5 is still down, but no l5_shift key action was called,
and current level is not 5, but 1.

Prevent this by not using the when-modifier-is-in-effect VK table for
keys that bind to key actions.