fix dead keys producing lowercased letters
As of Windows 7, ToUnicode always produces lowercased symbols.
The issue is similar to capitalized letters of MapVirtualKey before.

Use VkKeyScan in a loop for all possible WCHAR values instead.

To avoid doing it for every VK in a layout, pre-compute the values for
while compiling a KLC, calling it "cache".